requirements.txt
django
gunicorn==20.0.4
fastapi==0.66.0
uvicorn==0.14.0
python-decouple==3.4
psycopg2-binary==2.9.1
kafka-python
black
